requirements.txt
absl-py
aiohttp
aiosignal
aitextgen==0.6.0
apscheduler
async-timeout
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
filelock
fire
Flask
Flask-Login
Flask-Mail
Flask-pyWebPush
Flask-SQLAlchemy
frozenlist
fsspec
google-auth
google-auth-oauthlib
greenlet
grpcio
http-ece
huggingface-hub
idna
importlib-metadata
itsdangerous
Jinja2
lightning-utilities
lxml
Markdown
MarkupSafe
multidict
numpy
oauthlib
packaging
pandas
protobuf
py-vapid
pyasn1
pyasn1-modules
pycparser
pyDeprecate
python-dateutil
python-dotenv
pytorch-lightning==1.7.7
pytrends
pytz
pywebpush
PyYAML
regex
requests
requests-oauthlib
rsa
six
SQLAlchemy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
tokenizers
torch
torchmetrics
tqdm
transformers
typeguard
typing_extensions
urllib3
Werkzeug
yarl
zipp
zope.interface
